requirements.txt
pywebview>=6.1
PyQt5>=5.15.11; platform_system!="Windows"
PyQtWebEngine>=5.15.7; platform_system!="Windows"
pyyaml>=6.0.3
watchdog>=2.3
rich>=13.0
pynput>=1.7
psutil>=5.9
pystray>=0.19.5
Pillow>=10.0.0
